+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueEffect.kt
@@ -29,3 +29,7 @@ data class ShowOverdueAppointments(
 ) : OverdueEffect()
 
 object ShowNoActiveNetworkConnectionDialog : OverdueEffect()
+
+object OpenSelectDownloadFormatDialog : OverdueEffect()
+
+object OpenSelectShareFormatDialog : OverdueEffect()