+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/SelectOverdueDownloadFormatEffect.kt
@@ -1,16 +1,16 @@
 package org.simple.clinic.overdue.download.formatdialog
 
 import android.net.Uri
-import org.simple.clinic.overdue.download.OverdueListDownloadFormat
+import org.simple.clinic.overdue.download.OverdueListFileFormat
 
 sealed class SelectOverdueDownloadFormatEffect
 
-data class DownloadForShare(val downloadFormat: OverdueListDownloadFormat) : SelectOverdueDownloadFormatEffect()
+data class DownloadForShare(val fileFormat: OverdueListFileFormat) : SelectOverdueDownloadFormatEffect()
 
-data class ScheduleDownload(val downloadFormat: OverdueListDownloadFormat) : SelectOverdueDownloadFormatEffect()
+data class ScheduleDownload(val fileFormat: OverdueListFileFormat) : SelectOverdueDownloadFormatEffect()
 
 sealed class SelectOverdueDownloadFormatViewEffect : SelectOverdueDownloadFormatEffect()
 
-data class ShareDownloadedFile(val uri: Uri) : SelectOverdueDownloadFormatViewEffect()
+data class ShareDownloadedFile(val uri: Uri, val mimeType: String) : SelectOverdueDownloadFormatViewEffect()
 
 object Dismiss : SelectOverdueDownloadFormatViewEffect()