+++ app/src/test/java/org/simple/clinic/home/overdue/OverdueUpdateTest.kt
@@ -8,6 +8,7 @@ import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
 import org.simple.clinic.TestData
+import org.simple.clinic.analytics.NetworkConnectivityStatus.ACTIVE
 import org.simple.clinic.analytics.NetworkConnectivityStatus.INACTIVE
 import org.simple.clinic.facility.FacilityConfig
 import java.time.LocalDate
@@ -90,6 +91,17 @@ class OverdueUpdateTest {
   }
 
   @Test
+  fun `when download overdue list button is clicked and network is connected, then open select download format dialog`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(DownloadOverdueListClicked(networkStatus = Optional.of(ACTIVE)))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(OpenSelectDownloadFormatDialog)
+        ))
+  }
+
+  @Test
   fun `when share overdue list button is clicked and network is not connected, then show no active connection dialog`() {
     updateSpec
         .given(defaultModel)
@@ -97,6 +109,17 @@ class OverdueUpdateTest {
         .then(assertThatNext(
             hasNoModel(),
             hasEffects(ShowNoActiveNetworkConnectionDialog)
+        ))
+  }
+
+  @Test
+  fun `when share overdue list button is clicked and network is connected, then open select share format dialog`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(ShareOverdueListClicked(networkStatus = Optional.of(ACTIVE)))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(OpenSelectShareFormatDialog)
         ))
   }
 }